# openapi.yaml
openapi: 3.0.3
info:
  title: WolfGrid API
  description: |
    REST API for the WolfGrid web application — a field prospecting platform for real estate agents.

    ## Important notes for contributors

    ### iOS coupling
    The iOS app (FLYR) calls many of these routes directly in production. Routes tagged
    with `x-ios-coupled: true` must not be modified without coordinating with iOS development.
    See ARCHITECTURE.md Section 10 for the full coupling map.

    ### Route status
    Each route has an `x-status` extension with one of the following values:
    - `working` — verified functional
    - `broken` — known to return incorrect results or errors
    - `stub` — exists but returns hardcoded/empty response
    - `unknown` — not verified

    ### Authentication
    Most routes require an authenticated Supabase session passed via cookies.
    iOS routes use a Bearer token in the Authorization header.
    Admin routes require the service role key or a founder-level entitlement.

  version: "1.0.0"
  contact:
    name: Harry Brown
servers:
  - url: https://wolfgrid.app
    description: Production
  - url: https://wolfgrid.app
    description: Production (www)
  - url: http://localhost:3000
    description: Local development
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sb-access-token
      description: Supabase session cookie (web)
    bearerAuth:
      type: http
      scheme: bearer
      description: Supabase JWT (iOS)
...
paths:
...
  /api/integrations/test-lead:
    post:
      tags: [integrations]
      summary: Send test lead across all active integrations
      x-status: working
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Test leads sent

  # ── FUB ALIASES (iOS compatibility) ──────────────────────────
  # These routes are aliases of the followupboss routes above.
  # They exist specifically for iOS compatibility and must not be removed.
...